// src/app/core/api/model/instanceResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstanceDto } from './instanceDto';
import { ProviderErrorDto } from './providerErrorDto';


export interface InstanceResponseDto { 
    /**
     * List of instances
     */
    data: Array<InstanceDto>;
    /**
     * Partial errors from providers
     */
    partialErrors?: Array<ProviderErrorDto>;
}