// src/app/core/api/model/providerErrorDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProviderErrorDto { 
    /**
     * Provider identifier
     */
    provider: string;
    /**
     * Error message
     */
    message: string;
}